+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ---
 
+## [v1.1.0](https://github.com/phil1436/VSToolKit/tree/1.1.0) (2023-11-24)
+
+-   Command `Add Markdown Overview` added
+-   Color theme `VSToolKit Theme dark` added
+
+---
+
 ## [v1.0.3](https://github.com/phil1436/VSToolKit/tree/1.0.3) (2023-7-19)
 
 -   Bug fixes